netbox-pyats-webinar/requirements.txt
python-dotenv
pyats[library]
pyats.contrib